Modified Files: 	interbase/jrd/cmp.c
10/13/2000

Fix for bug #116579; Generators in computed by columns
will return wrong results.

If, in an ODS 10 or greater database, the gen_id function is
included as a part of a computed field and another field
references it or it is referenced in a select statement, the
database will become corrupted.  This problem was introduced with
the change to 64 bit generator values.

The COMPUTED BY clause is being parsed correctly, correct BLR
is stored in the database, and this BLR is correctly converted
into execution nodes by jrd/evl.c

During jrd/cmp.c  'copy'  processing, called as a part of
the pass1 routine, this execution node is modified replacing
the correct generator number with incorrect values.

These changes fix that problem by causing the correct
node-copying routine to be invoked.

pat