Please consider this contrived example:
create table t0 (
 a integer,
 genid_field1 computed by (gen_id(gen1, 1)),
 genid_field2 computed by (gen_id(gen2, genid_field1)),
 genid_field3 computed by (gen_id(gen3, genid_field2));

While I don't consider this to be an example of real-word SQL
coding, it does show that the second argument of the gen_id
function can be something other than a simple constant.

As jrd/cmp.c is coded now this statement will get processed
properly, but there is a serious problem:  The routine 'pass1'
will attempt to parse the internal generator ID as if it was
a pointer to an execution node.

This change corrects 'pass1' processing to only pass the
second argument on for further processing.


pat