Changed FIRST/SKIP syntax to:
SELECT [FIRST X | FIRST ? | FIRST (value)] [SKIP X | SKIP ? | SKIP (value)] ...
where X is a long integer ONLY.  The syntax still accepts full value
expressions, but it requires the parens around them.

because with the old syntax and the following query:

SELECT FIRST 5 * FROM MESSAGES;

the parser was treating the * as the times symbol and not the field list.
This caused an error because '5 * FROM' doesn't parse.